+++ app/entities/UserEntity.php
@@ -9,9 +9,8 @@ namespace App\Entities;
  */
 class UserEntity extends AEntity {
     private string $id;
-    private string $username;
     private string $fullname;
-    private ?string $email;
+    private string $email;
     private string $dateCreated;
     private bool $isTechnical;
     private int $appDesignTheme;
@@ -19,35 +18,44 @@ class UserEntity extends AEntity {
     private string $timeFormat;
     private ?string $superiorUserId;
     private ?string $profilePictureFileId;
+    private ?string $orgPosition;
+    private ?string $orgDepartment;
+    private ?string $orgSection;
+    private ?string $personalNumber;
 
     /**
      * Class constructor
      * 
      * @param string $id User ID
-     * @param string $username Username
      * @param string $fullname Fullname
-     * @param ?string $email Email
+     * @param string $email Email
      * @param string $dateCreated Date created
      * @param bool $isTechnical Is user technical
      * @param int $appDesignTheme App design theme
      * @param ?string $superiorUserId Superior user ID
      * @param ?string $profilePictureFileId Profile picture file ID
+     * @param ?string $orgPosition Organization position
+     * @param ?string $orgDepartment Organization department
+     * @param ?string $orgSection Organization section
+     * @param ?string $personalNumber Personal number
      */
     public function __construct(
         string $id,
-        string $username,
         string $fullname,
-        ?string $email,
+        string $email,
         string $dateCreated,
         bool $isTechnical,
         int $appDesignTheme,
         string $dateFormat,
         string $timeFormat,
         ?string $superiorUserId,
-        ?string $profilePictureFileId
+        ?string $profilePictureFileId,
+        ?string $orgPosition,
+        ?string $orgDepartment,
+        ?string $orgSection,
+        ?string $personalNumber
     ) {
         $this->id = $id;
-        $this->username = $username;
         $this->fullname = $fullname;
         $this->email = $email;
         $this->dateCreated = $dateCreated;
@@ -57,6 +65,10 @@ class UserEntity extends AEntity {
         $this->timeFormat = $timeFormat;
         $this->superiorUserId = $superiorUserId;
         $this->profilePictureFileId = $profilePictureFileId;
+        $this->orgPosition = $orgPosition;
+        $this->orgDepartment = $orgDepartment;
+        $this->orgSection = $orgSection;
+        $this->personalNumber = $personalNumber;
     }
 
     /**
@@ -64,13 +76,6 @@ class UserEntity extends AEntity {
      */
     public function getId(): string {
         return $this->id;
-    }
-
-    /**
-     * Returns username
-     */
-    public function getUsername(): string {
-        return $this->username;
     }
 
     /**
@@ -83,7 +88,7 @@ class UserEntity extends AEntity {
     /**
      * Returns email
      */
-    public function getEmail(): ?string {
+    public function getEmail(): string {
         return $this->email;
     }
 
@@ -162,18 +167,46 @@ class UserEntity extends AEntity {
         return $initials;
     }
 
+    /**
+     * Returns user's position in organization
+     */
+    public function getOrgPosition(): ?string {
+        return $this->orgPosition;
+    }
+
+    /**
+     * Returns user's section in organization
+     */
+    public function getOrgSection(): ?string {
+        return $this->orgSection;
+    }
+
+    /**
+     * Returns user's department in organization
+     */
+    public function getOrgDepartment(): ?string {
+        return $this->orgDepartment;
+    }
+
+    /**
+     * Returns user's personal number
+     */
+    public function getPersonalNumber(): ?string {
+        return $this->personalNumber;
+    }
+
     public static function createEntityFromDbRow(mixed $row): ?static {
         if($row === null) {
             return null;
         }
 
         $row = self::createRow($row);
-        self::checkTypes($row, ['userId' => 'string', 'username' => 'string', 'fullname' => 'string', 'email' => '?string', 'dateCreated' => 'string', 'isTechnical' => 'bool', 'appDesignTheme' => 'int',
-                                'dateFormat' => 'string', 'timeFormat' => 'string', 'superiorUserId' => '?string', 'profilePictureFileId' => '?string']);
+        self::checkTypes($row, ['userId' => 'string', 'fullname' => 'string', 'email' => 'string', 'dateCreated' => 'string', 'isTechnical' => 'bool', 'appDesignTheme' => 'int',
+                                'dateFormat' => 'string', 'timeFormat' => 'string', 'superiorUserId' => '?string', 'profilePictureFileId' => '?string', 'orgPosition' => '?string',
+                                'orgSection' => '?string', 'orgDepartment' => '?string', 'personalNumber' => '?string']);
 
         return new self(
             $row->userId,
-            $row->username,
             $row->fullname,
             $row->email,
             $row->dateCreated,
@@ -182,7 +215,11 @@ class UserEntity extends AEntity {
             $row->dateFormat,
             $row->timeFormat,
             $row->superiorUserId,
-            $row->profilePictureFileId
+            $row->profilePictureFileId,
+            $row->orgPosition,
+            $row->orgDepartment,
+            $row->orgSection,
+            $row->personalNumber
         );
     }
 }