+++ app/modules/AdminModule/Presenters/Members/UsersPresenter.php
@@ -7,6 +7,8 @@ use App\Core\AjaxRequestBuilder;
 use App\Core\Caching\CacheNames;
 use App\Core\DB\DatabaseRow;
 use App\Core\HashManager;
+use App\Core\Http\Ajax\Operations\CustomOperation;
+use App\Core\Http\Ajax\Requests\PostAjaxRequest;
 use App\Core\Http\FormRequest;
 use App\Core\Http\HttpRequest;
 use App\Core\Http\JsonResponse;
@@ -39,13 +41,12 @@ class UsersPresenter extends AAdminPresenter {
         $grid->createDataSourceFromQueryBuilder($qb, 'userId');
 
         $grid->addColumnText('fullname', 'Fullname');
-        $grid->addColumnText('username', 'Username');
         $grid->addColumnText('email', 'Email');
         $grid->addColumnBoolean('isTechnical', 'Technical user');
         $grid->addColumnBoolean('isDeleted', 'Is deleted');
 
         $grid->addQuickSearch('fullname', 'Fullname');
-        $grid->addQuickSearch('username', 'Username');
+        $grid->addQuickSearch('email', 'Email');
 
         $grid->addFilter('isTechnical', 0, ['0' => 'False', '1' => 'True']);
         $grid->addFilter('isDeleted', 0, ['0' => 'False', '1' => 'True']);
@@ -110,23 +111,39 @@ class UsersPresenter extends AAdminPresenter {
             // Add user to container All users group
 
             try {
+                // USER CREATION
                 $this->app->userRepository->beginTransaction(__METHOD__);
 
-                $email = $fr->email;
-                if($email == '') {
-                    $email = null;
-                }
-
-                $userId = $this->app->userManager->createNewUser($fr->username, $fr->fullname, HashManager::hashPassword($fr->password), $email);
+                $userId = $this->app->userManager->createNewUser($fr->email, $fr->fullname, HashManager::hashPassword($fr->password));
 
                 $this->app->userRepository->commit($this->getUserId(), __METHOD__);
 
+                // USER UPDATE
                 $this->app->userRepository->beginTransaction(__METHOD__);
 
+                $updateData = [];
+
                 if($fr->superiorUser != 'null') {
-                    $this->app->userManager->updateUser($userId, ['superiorUserId' => $fr->superiorUser]);
+                    $updateData['superiorUserid'] = $fr->superiorUser;
+                }
+                if($fr->orgPosition !== null) {
+                    $updateData['orgPosition'] = $fr->orgPosition;
+                }
+                if($fr->orgDepartment !== null) {
+                    $updateData['orgDepartment'] = $fr->orgDepartment;
+                }
+                if($fr->orgSection !== null) {
+                    $updateData['orgSection'] = $fr->orgSection;
+                }
+                if($fr->personalNumber !== null) {
+                    $updateData['personalNumber'] = $fr->personalNumber;
                 }
 
+                if(!empty($updateData)) {
+                    $this->app->userManager->updateUser($userId, $updateData);
+                }
+
+                // ADD USER TO CONTAINER GROUP
                 $containerId = $this->httpSessionGet('container');
                 $container = $this->app->containerManager->getContainerById($containerId);
 
@@ -135,6 +152,7 @@ class UsersPresenter extends AAdminPresenter {
 
                 $this->app->groupManager->addUserToGroup($userId, $group->groupId);
 
+                // ADD USER TO ALL USERS IN CONTAINER
                 $this->groupRepository->beginTransaction(__METHOD__);
                 
                 $this->groupManager->addUserToGroupTitle(SystemGroups::ALL_USERS, $userId);
@@ -167,7 +185,7 @@ class UsersPresenter extends AAdminPresenter {
 
         $form->setAction($this->createURL('newUserForm'));
 
-        $form->addTextInput('username', 'Username:')
+        $form->addEmailInput('email', 'Email:')
             ->setRequired();
 
         $form->addTextInput('fullname', 'Fullname:')
@@ -175,8 +193,6 @@ class UsersPresenter extends AAdminPresenter {
 
         $form->addPasswordInput('password', 'Password:')
             ->setRequired();
-
-        $form->addEmailInput('email', 'Email:');
 
         $form->addHorizontalLine();
 
@@ -186,6 +202,13 @@ class UsersPresenter extends AAdminPresenter {
 
         $form->addSelect('superiorUser', 'Superior user:')
             ->addRawOption('null', 'Not selected', true);
+
+        $form->addHorizontalLine();
+
+        $form->addTextInput('orgPosition', 'Position:');
+        $form->addTextInput('orgSection', 'Section:');
+        $form->addTextInput('orgDepartment', 'Department:');
+        $form->addTextInput('personalNumber', 'Personal number:');
 
         $form->addSubmit('Create');
 
@@ -214,7 +237,51 @@ class UsersPresenter extends AAdminPresenter {
             }
         ');
 
+        $par = new PostAjaxRequest($this->httpRequest);
+        $par->setUrl($this->createURL('checkEmailForNewUserForm'))
+            ->addArgument('_query')
+            ->setData(['query' => '_query']);
+
+        $operation = new CustomOperation();
+
+        $operation->addCode('
+            if(obj.result == 0) {
+                alert("This email is taken.");
+                $("#formSubmit").attr("disabled", true);
+                $("#email").css("border", "1px solid red");
+            } else {
+                $("#formSubmit").removeAttr("disabled");
+                $("#email").css("border", "1px solid black");
+            }
+        ');
+
+        $par->addOnFinishOperation($operation);
+
+        $form->addScript($par);
+
+        $form->addScript('
+            async function checkEmailExists() {
+                const query = $("#email").val();
+
+                await ' . $par->getFunctionName() . '(query);
+            }
+
+            $("form").on("focusout", "#email", async function() {
+                await checkEmailExists();
+            });
+        ');
+
         return $form;
+    }
+
+    public function actionCheckEmailForNewUserForm() {
+        $query = $this->httpRequest->get('query');
+
+        $users = $this->app->userRepository->searchUsers($query, ['email']);
+
+        return new JsonResponse([
+            'result' => empty($users) ? 1 : 0
+        ]);
     }
 
     public function renderEditUserForm() {
@@ -260,6 +327,17 @@ class UsersPresenter extends AAdminPresenter {
             } catch(AException $e) {}
         }
 
+        $form->addHorizontalLine();
+
+        $form->addTextInput('orgPosition', 'Position:')
+            ->setValue($user->getOrgPosition());
+        $form->addTextInput('orgSection', 'Section:')
+            ->setValue($user->getOrgSection());
+        $form->addTextInput('orgDepartment', 'Department:')
+            ->setValue($user->getOrgDepartment());
+        $form->addTextInput('personalNumber', 'Personal number:')
+            ->setValue($user->getPersonalNumber());
+
         $form->addSubmit('Create');
 
         $arb = new AjaxRequestBuilder();
@@ -300,7 +378,11 @@ class UsersPresenter extends AAdminPresenter {
 
             $this->app->userManager->updateUser($userId, [
                 'fullname' => $fr->fullname,
-                'superiorUserId' => $fr->superiorUser
+                'superiorUserId' => $fr->superiorUser,
+                'orgPosition' => $fr->orgPosition,
+                'orgSection' => $fr->orgSection,
+                'orgDepartment' => $fr->orgDepartment,
+                'personalNumber' => $fr->personalNumber
             ]);
 
             $this->app->userRepository->commit($this->getUserId(), __METHOD__);
@@ -358,23 +440,38 @@ class UsersPresenter extends AAdminPresenter {
     public function actionSearchUsers() {
         $query = $this->httpRequest->get('query');
 
-        $users = [
-            '<option value="null">Not selected</option>'
-        ];
+        $users = [];
+
+        if(!$this->isSuperiorUserMandatory()) {
+            $users[] = '<option value="null">Not selected</option>';
+        }
 
         $container = $this->app->containerManager->getContainerById($this->containerId);
 
         $userIds = $this->app->groupManager->getGroupUsersForGroupTitle($container->getTitle() . ' - users');
 
-        foreach($userIds as $userId) {
-            $user = $this->app->userManager->getUserById($userId);
+        $qb = $this->app->userRepository->composeQueryForUsers();
 
-            if(str_contains($user->getFullname(), $query) || str_contains($user->getUsername(), $query) || str_contains($user->getEmail(), $query)) {
-                $users[] = '<option value="' . $user->getId() . '">' . $user->getFullname() . '</option>';
-            }
+        $qb->andWhere($qb->getColumnInValues('userId', $userIds))
+            ->andWhere('(fullname LIKE ? OR email LIKE ?)', ['%' . $query . '%', '%' . $query . '%'])
+            ->execute();
+
+        while($row = $qb->fetchAssoc()) {
+            $users[] = '<option value="' . $row['userId'] . '">' . $row['fullname'] . '</option>';
         }
 
         return new JsonResponse(['users' => $users]);
+    }
+
+    /**
+     * Returns true if superior user must be selected during new user creation
+     */
+    private function isSuperiorUserMandatory(): bool {
+        $container = $this->app->containerManager->getContainerById($this->containerId);
+
+        $userIds = $this->app->groupManager->getGroupUsersForGroupTitle($container->getTitle() . ' - users');
+
+        return count($userIds) > 1;
     }
 }